part 1 passes 100%

+++ ColorMapInput.c
@@ -23,13 +23,13 @@ uint8_t** FileToColorMap(char* colorfile, int* colorcount)
     colorcount = &count; 
     fscanf(file, "%d", &count);
     uint8_t** result = (uint8_t**) malloc(*colorcount * sizeof(uint8_t*));  
-    for (uint8_t i = 0; i < *colorcount; i++) {
+    for (int i = 0; i < *colorcount; i++) {
 	uint8_t* color = (uint8_t*) malloc(3 * sizeof(uint8_t));  
 	fscanf(file, "%hhu", &color[0]);
 	fscanf(file, "%hhu", &color[1]); 
-	fscanf(file, "%hhu", &color[2]); 
- 	result[i] = color; 
-        }
+	fscanf(file, "%hhu", &color[2]);  
+	result[i] = color;
+        }  
     return result;
 }
